refactor(docs): relocate tooling files
