+++ src/styleguide/vueDocsBlockPlugin.js
