+++ vite.config.ts
@@ -7,7 +7,7 @@ import { createAppConfig } from '@nextcloud/vite-config'
 import eslint from 'vite-plugin-eslint'
 import stylelint from 'vite-plugin-stylelint'
 import path from 'path'
-import { vueDocsBlockPlugin } from './tools/vueDocsBlockPlugin.js'
+import { vueDocsBlockPlugin } from './src/styleguide/vueDocsBlockPlugin.js'
 
 const isProduction = process.env.NODE_ENV === 'production'
